refactor: moved methods for creating reporter to the Reporter
