+++ src/runner/bootstrapper.ts
@@ -1,11 +1,8 @@
-import path from 'path';
-import fs from 'fs';
 import {
     chunk,
     times,
 } from 'lodash';
 
-import makeDir from 'make-dir';
 import debug from 'debug';
 import prettyTime from 'pretty-hrtime';
 import Compiler from '../compiler';
@@ -15,18 +12,15 @@ import { GeneralError } from '../errors/runtime';
 import { RUNTIME_ERRORS } from '../errors/types';
 import TestedApp from './tested-app';
 import parseFileList from '../utils/parse-file-list';
-import resolvePathRelativelyCwd from '../utils/resolve-path-relatively-cwd';
 import loadClientScripts from '../custom-client-scripts/load';
 import { getConcatenatedValuesString } from '../utils/string';
-import { Writable as WritableStream } from 'stream';
-import { ReporterSource, ReporterPluginSource } from '../reporter/interfaces';
+import { ReporterSource } from '../reporter/interfaces';
 import ClientScript from '../custom-client-scripts/client-script';
 import ClientScriptInit from '../custom-client-scripts/client-script-init';
 import BrowserConnectionGateway from '../browser/connection/gateway';
 import { CompilerArguments } from '../compiler/interfaces';
 import CompilerService from '../services/compiler/host';
 import Test from '../api/structure/test';
-import { getPluginFactory, processReporterName } from '../utils/reporter';
 import { BootstrapperInit, BrowserSetOptions } from './interfaces';
 import WarningLog from '../notifications/warning-log';
 import WARNING_MESSAGES from '../notifications/warning-message';
@@ -55,7 +49,6 @@ interface BasicRuntimeResources {
 }
 
 interface RuntimeResources extends BasicRuntimeResources {
-    reporterPlugins: ReporterPluginSource[];
     commonClientScripts: ClientScript[];
 }
 
@@ -231,41 +224,6 @@ export default class Bootstrapper {
         return tests;
     }
 
-    private async _ensureOutStream (outStream: string | WritableStream): Promise<WritableStream> {
-        if (typeof outStream !== 'string')
-            return outStream;
-
-        const fullReporterOutputPath = resolvePathRelativelyCwd(outStream);
-
-        await makeDir(path.dirname(fullReporterOutputPath));
-
-        return fs.createWriteStream(fullReporterOutputPath);
-    }
-
-    private static _addDefaultReporter (reporters: ReporterSource[]): void {
-        reporters.push({
-            name:   'spec',
-            output: process.stdout,
-        });
-    }
-
-    private async _getReporterPlugins (): Promise<ReporterPluginSource[]> {
-        if (!this.reporters.length)
-            Bootstrapper._addDefaultReporter(this.reporters);
-
-        return Promise.all(this.reporters.map(async ({ name, output }) => {
-            const pluginFactory = getPluginFactory(name);
-            const processedName = processReporterName(name);
-            const outStream     = output ? await this._ensureOutStream(output) : void 0;
-
-            return {
-                plugin: pluginFactory(),
-                name:   processedName,
-                outStream,
-            };
-        }));
-    }
-
     private async _startTestedApp (): Promise<TestedApp|undefined> {
         if (!this.appCommand)
             return void 0;
@@ -355,12 +313,11 @@ export default class Bootstrapper {
 
     // API
     public async createRunnableConfiguration (): Promise<RuntimeResources> {
-        const reporterPlugins     = await this._getReporterPlugins();
         const commonClientScripts = await loadClientScripts(this.clientScripts);
 
         if (await this._canUseParallelBootstrapping(this.browsers))
-            return { reporterPlugins, ...await this._bootstrapParallel(this.browsers), commonClientScripts };
+            return { ...await this._bootstrapParallel(this.browsers), commonClientScripts };
 
-        return { reporterPlugins, ...await this._bootstrapSequence(this.browsers), commonClientScripts };
+        return { ...await this._bootstrapSequence(this.browsers), commonClientScripts };
     }
 }
